+++ src/tokens.ts
@@ -7,7 +7,7 @@ import * as crypto from "crypto"
 import exchangeCodeForTokensFactory from "./exchange-code-for-token"
 
 const random = (length = 32) =>
-  jose.base64url.encode(crypto.randomBytes(length))
+  jose.base64url.encode(new Uint8Array(crypto.randomBytes(length)))
 
 const createSignedJWT = async ({
   alg,